# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Security Agent Suite API
  version: 0.1.0
  description: >-
    Go control-plane API for five independently callable cybersecurity agents.
    All /v1 endpoints require X-Tenant-ID and, when configured, X-API-Key.
servers:
  - url: http://127.0.0.1:8080
security:
  - ApiKeyAuth: []
...
paths:
...
  /v1/agents:
    get:
      tags: [Agents]
      summary: List callable agents
      parameters:
        - $ref: "#/components/parameters/TenantID"
      responses:
        "200":
          description: Agent catalog
          content:
            application/json:
              schema:
                type: object
                required: [agents]
                properties:
                  agents:
                    type: array
                    items: { $ref: "#/components/schemas/Agent" }
        default: { $ref: "#/components/responses/Error" }
  /v1/agents/{agentID}:
    get:
      tags: [Agents]
      summary: Get one agent
      parameters:
        - $ref: "#/components/parameters/TenantID"
        - $ref: "#/components/parameters/AgentID"
      responses:
        "200":
          description: Agent definition
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Agent" }
        default: { $ref: "#/components/responses/Error" }
  /v1/agents/{agentID}/runs:
    post:
      tags: [Runs]
      summary: Create an agent run
      description: >-
        request_id is idempotent within a tenant. Active attack-path validation
        is created in waiting_approval and is not queued until approved.
        Opt-in execution_mode=manual is restricted to event-triage/triage with
        network_access=deny. It creates preparing (201), never schedules before
        explicit submit, and compares a versioned normalized creation fingerprint.
        Automatic creation remains 202; replay is 200. Unknown or missing historical
        manual fingerprints and automatic/manual request_id collisions are 409.
      parameters:
        - $ref: "#/components/parameters/TenantID"
        - $ref: "#/components/parameters/AgentID"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateRunRequest" }
      responses:
        "201":
          description: Manual Run prepared without scheduling
          headers:
            Location: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Run" }
        "202":
          description: Run created
          headers:
            Location: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Run" }
        "200":
          description: Existing idempotent run returned
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Run" }
        default: { $ref: "#/components/responses/Error" }
  /v1/runs:
    get:
      tags: [Runs]
      summary: List tenant runs
      parameters:
        - $ref: "#/components/parameters/TenantID"
        - name: agent_id
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { $ref: "#/components/schemas/RunStatus" }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
      responses:
        "200":
          description: Runs
          content:
            application/json:
              schema:
                type: object
                required: [runs, count]
                properties:
                  runs:
                    type: array
                    items: { $ref: "#/components/schemas/Run" }
                  count: { type: integer }
        default: { $ref: "#/components/responses/Error" }
  /v1/runs/{runID}:
    get:
      tags: [Runs]
      summary: Get a run
      parameters:
        - $ref: "#/components/parameters/TenantID"
        - $ref: "#/components/parameters/RunID"
      responses:
        "200":
          description: Run
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Run" }
        default: { $ref: "#/components/responses/Error" }
  /v1/runs/{runID}/events:
    get:
      tags: [Runs]
      summary: Get run audit events
      parameters:
        - $ref: "#/components/parameters/TenantID"
        - $ref: "#/components/parameters/RunID"
      responses:
        "200":
          description: Events
          content:
            application/json:
              schema:
                type: object
                required: [run_id, events]
                properties:
                  run_id: { type: string }
                  events:
                    type: array
                    items: { $ref: "#/components/schemas/RunEvent" }
        default: { $ref: "#/components/responses/Error" }
  /v1/runs/{runID}/evidence:
    post:
      tags: [Runs]
      summary: Append immutable run evidence
      description: >-
        Appends one immutable Evidence record. The server may generate the id
        and collection time when omitted; existing evidence ids cannot be
        replaced. When supplied, Artifact references must belong to the same
        Run and match the Evidence hash.
      parameters:
        - $ref: "#/components/parameters/TenantID"
        - $ref: "#/components/parameters/RunID"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EvidenceAppendRequest" }
      responses:
        "201":
          description: Evidence appended
          headers:
            Location: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StoredEvidence" }
        default: { $ref: "#/components/responses/Error" }
    get:
      tags: [Runs]
      summary: List immutable run evidence
      parameters:
        - $ref: "#/components/parameters/TenantID"
        - $ref: "#/components/parameters/RunID"
      responses:
        "200":
          description: Evidence records in deterministic collection order
          content:
            application/json:
              schema:
                type: object
                required: [run_id, evidence]
                properties:
                  run_id: { type: string }
                  evidence:
                    type: array
                    items: { $ref: "#/components/schemas/StoredEvidence" }
        default: { $ref: "#/components/responses/Error" }
  /v1/runs/{runID}/evidence/{evidenceID}:
    get:
      tags: [Runs]
      summary: Get one immutable Evidence record
      parameters:
        - $ref: "#/components/parameters/TenantID"
        - $ref: "#/components/parameters/RunID"
        - name: evidenceID
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Evidence record
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StoredEvidence" }
        default: { $ref: "#/components/responses/Error" }
  /v1/runs/{runID}/artifacts:
    post:
      tags: [Runs]
      summary: Upload and register a run artifact
      description: >-
        Streams the request body into the run-scoped artifact store, computes
        SHA-256, and returns metadata rather than embedding bytes in the Run.
        X-Artifact-SHA256 is optional; when supplied, a mismatch is rejected.
        Manual Runs only accept uploads while preparing, require X-Artifact-SHA256
        to match the frozen manifest, and accept application/json with optional
        charset=utf-8. Actual bytes must match digest and size, bounded by the lower
        of 1 MiB and the server body limit. Upload never submits. A post-publication
        500 can leave a registered Artifact; query before retrying.
      parameters:
        - $ref: "#/components/parameters/TenantID"
        - $ref: "#/components/parameters/RunID"
        - name: name
          in: query
          required: true
          schema: { type: string, minLength: 1 }
        - name: X-Artifact-SHA256
          in: header
          required: false
          schema: { type: string, pattern: "^[0-9a-fA-F]{64}$" }
      requestBody:
        required: true
        description: Raw artifact bytes. Content-Type is stored as artifact metadata.
        content:
          "*/*":
            schema: { type: string, format: binary }
      responses:
        "201":
          description: Artifact stored and registered
          headers:
            Location: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Artifact" }
        "413":
          $ref: "#/components/responses/Error"
        default: { $ref: "#/components/responses/Error" }
    get:
      tags: [Runs]
      summary: List run artifacts
      parameters:
        - $ref: "#/components/parameters/TenantID"
        - $ref: "#/components/parameters/RunID"
      responses:
        "200":
          description: Artifact metadata
          content:
            application/json:
              schema:
                type: object
                required: [run_id, artifacts]
                properties:
                  run_id: { type: string }
                  artifacts:
                    type: array
                    items: { $ref: "#/components/schemas/Artifact" }
        default: { $ref: "#/components/responses/Error" }
  /v1/runs/{runID}/artifacts/{artifactID}:
    get:
      tags: [Runs]
      summary: Download one run artifact
      parameters:
        - $ref: "#/components/parameters/TenantID"
        - $ref: "#/components/parameters/RunID"
        - name: artifactID
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Artifact bytes
          content:
            application/json:
              schema: {}
            text/plain:
              schema: { type: string }
            application/octet-stream:
              schema: { type: string, format: binary }
        default: { $ref: "#/components/responses/Error" }
  /v1/runs/{runID}/submit:
    post:
      tags: [Runs]
      summary: Explicitly submit one validated manual input
      description: >-
        Strict, case-sensitive single JSON object; rejects duplicate keys, null,
        invalid UTF-8 and unpaired surrogates. Reads same-Run Artifact bytes and
        validates SHA-256, size, media type and sas.synthetic-alert/v1 before
        durable Evidence and frozen input publication. Same Artifact replay is
        200 even after completion; another Artifact or cancelled preparation is
        409. Other-tenant/Run resources are 404. Invalid bytes/schema are 400;
        input too large is 413 artifact_too_large and control body too large is
        413 request_too_large. Internal errors are 500 with uncertain persistence;
        query/retry the same identifiers, never create a new execution to retry.
      parameters:
        - $ref: "#/components/parameters/TenantID"
        - $ref: "#/components/parameters/RunID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [artifact_id]
              properties:
                artifact_id: { type: string, minLength: 1, maxLength: 128, pattern: '^[A-Za-z0-9._-]+$' }
      responses:
        "202":
          description: Frozen input durably accepted; queue notification is recoverable
          headers:
            Location: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Run" }
        "200":
          description: Same submission replay; no additional execution
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Run" }
        default: { $ref: "#/components/responses/Error" }
  /v1/runs/{runID}/approve:
    post:
      tags: [Runs]
      summary: Approve a waiting run
      parameters:
        - $ref: "#/components/parameters/TenantID"
        - $ref: "#/components/parameters/RunID"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ApprovalRequest" }
      responses:
        "202":
          description: Approved and queued
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Run" }
        default: { $ref: "#/components/responses/Error" }
  /v1/runs/{runID}/cancel:
    post:
      tags: [Runs]
      summary: Cancel a queued or running run
      parameters:
        - $ref: "#/components/parameters/TenantID"
        - $ref: "#/components/parameters/RunID"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                actor: { type: string }
                reason: { type: string }
      responses:
        "200":
          description: >-
            Cancellation has already reached a terminal Run. For a running
            API-requested cancellation, the terminal result uses
            `executor_cancelled`; a runtime canceled envelope under an
            unexpired Go context uses `agent_compose_cancelled`.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Run" }
        "202":
          description: >-
            Cancellation requested for a running Run; stopping is still in
            progress. The authoritative Go context eventually records terminal
            `cancelled` with `error_code=executor_cancelled`; poll the Run until
            terminal. Successfully decoded runtime provenance remains available.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Run" }
        default: { $ref: "#/components/responses/Error" }
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    TenantID:
      name: X-Tenant-ID
      in: header
      required: true
      schema: { type: string, minLength: 1, maxLength: 128 }
    AgentID:
      name: agentID
      in: path
      required: true
      schema:
        type: string
        enum: [traffic-analysis, event-triage, attack-path-validation, compliance-query, security-report]
    RunID:
      name: runID
      in: path
      required: true
      schema: { type: string }
  responses:
    Error:
      description: Structured error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            request_id: { type: string }
    Agent:
      type: object
      required: [id, display_name, description, risk, executor_agent, allowed_modes, input_types, output_formats, capset_ids, default_timeout]
      properties:
        id: { type: string }
        display_name: { type: string }
        description: { type: string }
        risk: { enum: [low, medium, high] }
        executor_agent: { type: string }
        allowed_modes: { type: array, items: { type: string } }
        input_types: { type: array, items: { type: string } }
        output_formats: { type: array, items: { type: string } }
        capset_ids: { type: array, items: { type: string } }
        approval_required_modes: { type: array, items: { type: string } }
        default_timeout: { type: string }
    RunStatus:
      type: string
      enum: [preparing, queued, validating, waiting_approval, running, succeeded, partial, failed, cancelled]
    InputRef:
      type: object
      additionalProperties: false
      required: [type, uri]
      properties:
        type: { type: string }
        uri: { type: string, format: uri }
        sha256: { type: string, pattern: "^[a-fA-F0-9]{64}$" }
        media_type: { type: string }
        size_bytes: { type: integer, minimum: 0 }
        metadata: { type: object, additionalProperties: { type: string } }
...
    Policy:
      type: object
      additionalProperties: false
      properties:
        network_access: { enum: [deny, restricted, allow] }
        active_validation: { type: boolean }
        approval_id: { type: string }
        max_requests: { type: integer, minimum: 0, maximum: 1000 }
        max_duration: { type: string }
...
    CreateRunRequest:
      $ref: "../contracts/run-request.schema.json"
    InputManifest:
      $ref: "../contracts/run-request.schema.json#/$defs/InputManifest"
    Submission:
      type: object
      additionalProperties: false
      required: [run_id, artifact_id, schema, sha256, size_bytes, media_type, evidence_id, submitted_at, fingerprint]
      properties:
        run_id: { type: string }
        artifact_id: { type: string }
        schema: { const: "sas.synthetic-alert/v1" }
        sha256: { type: string, pattern: "^[a-f0-9]{64}$" }
        size_bytes: { type: integer, minimum: 1, maximum: 1048576 }
        media_type: { const: application/json }
        evidence_id: { type: string }
        submitted_at: { type: string, format: date-time }
        fingerprint: { type: string, pattern: '^sas\.manual-submit/v1:[a-f0-9]{64}$' }
    ApprovalRequest:
      type: object
      additionalProperties: false
      required: [approval_id, actor, reason]
      properties:
        approval_id: { type: string, minLength: 1 }
        actor: { type: string, minLength: 1 }
        reason: { type: string, minLength: 1 }
...
    Artifact:
      type: object
      required: [id, name, uri, media_type, created_at]
      properties:
        id: { type: string }
        name: { type: string }
        uri: { type: string }
        media_type: { type: string }
        sha256: { type: string }
        size_bytes: { type: integer }
        created_at: { type: string, format: date-time }
...
    StoredEvidence:
      allOf:
        - $ref: "#/components/schemas/Evidence"
        - type: object
          required: [id, type, source_uri, sha256, tool, tool_version, parameters, collected_at]
    EvidenceAppendRequest:
      type: object
      additionalProperties: false
      required: [type, source_uri, sha256, tool, tool_version, parameters]
      properties:
        id: { type: string, minLength: 1 }
        type: { type: string, minLength: 1 }
        source_uri: { type: string, minLength: 1 }
        sha256: { type: string, pattern: "^[a-fA-F0-9]{64}$" }
        excerpt: { type: string }
        tool: { type: string, minLength: 1 }
        tool_version: { type: string, minLength: 1 }
        parameters: { type: object, additionalProperties: { type: string } }
        artifact_ids: { type: array, uniqueItems: true, items: { type: string, minLength: 1 } }
        collected_at: { type: string, format: date-time }
        metadata: { type: object, additionalProperties: { type: string } }
...
    RunEvent:
      type: object
      required: [sequence, type, message, at]
      properties:
        sequence: { type: integer }
        type: { type: string }
        from: { $ref: "#/components/schemas/RunStatus" }
        to: { $ref: "#/components/schemas/RunStatus" }
        message: { type: string }
        actor: { type: string }
        at: { type: string, format: date-time }
        details: { type: object, additionalProperties: { type: string } }
    Run:
      type: object
      required: [id, request_id, agent_id, agent_name, mode, scope, status, created_at, updated_at, version]
      properties:
        execution_mode: { type: string, enum: [automatic, manual] }
        input_manifest: { $ref: "#/components/schemas/InputManifest" }
        creation_fingerprint: { type: string, pattern: '^sas\.manual-create/v1:[a-f0-9]{64}$' }
        submission: { $ref: "#/components/schemas/Submission" }
        id: { type: string }
        request_id: { type: string }
        case_id: { type: string }
        agent_id: { type: string }
        agent_name: { type: string }
        mode: { type: string }
        inputs: { type: array, items: { $ref: "#/components/schemas/InputRef" } }
        scope: { $ref: "#/components/schemas/Scope" }
        policy: { $ref: "#/components/schemas/Policy" }
        output: { $ref: "#/components/schemas/OutputRequest" }
        metadata: { type: object, additionalProperties: { type: string } }
        status: { $ref: "#/components/schemas/RunStatus" }
        approval: { $ref: "#/components/schemas/Approval" }
        result: { $ref: "#/components/schemas/RunResult" }
        events: { type: array, items: { $ref: "#/components/schemas/RunEvent" } }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        started_at: { type: string, format: date-time }
        completed_at: { type: string, format: date-time }
        version: { type: integer }